requirements.txt
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.0.0
requests>=2.28.0
PySocks>=1.7.1
